don't split up top tier page rendering between layout and page
