+++ src/lib/components/Breadcrumbs/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./Breadcrumbs.svelte";
+export type { Breadcrumb, Breadcrumbs as BreadcrumbsType } from "./types";
