+++ src/lib/components/Breadcrumbs/types.ts
@@ -1,0 +1,7 @@
+export type Breadcrumb = {
+  id: string;
+  title: string | null | undefined;
+  link: string | null | undefined;
+};
+
+export type Breadcrumbs = Breadcrumb[];
