+++ src/routes/(infoPages)/[topTierPage]/page.scss
@@ -1,3 +1,7 @@
+.top-tier-page-hero-image {
+  height: 33vh;
+}
+
 // TODO: [LDAF-370] Create a <CardGroup /> Wrapper to
 // avoid dealing with  the <ul> and styling overrides here.
 .usa-prose .service-group-list {
