--- src/routes/(infoPages)/layout.scss
@@ -1,3 +1,0 @@
-.top-tier-page-hero-image {
-  height: 33vh;
-}
