+++ src/stories/Breadcrumbs.stories.ts
@@ -20,6 +20,6 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    path: breadcrumbsTestContent,
+    breadcrumbs: breadcrumbsTestContent,
   },
 };
